+++ frontend/src/components/forms_display/FormsList.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { useSelector, useDispatch } from "react-redux";
 import { getSurveys, reset } from "../../features/surveys/surveySlice";
 import Spinner from "../Spinner";
-import SurveyItem from "../SurveyItem";
+import SurveyItem from "./SurveyItem";
 import survey from "../../images/survey.png";
 import { Button } from "@chakra-ui/react";
