+++ frontend/src/components/forms_display/SurveyItem.jsx
@@ -1,5 +1,5 @@
 import { useDispatch } from "react-redux";
-import { deleteSurvey } from "../features/surveys/surveySlice";
+import { deleteSurvey } from "../../features/surveys/surveySlice";
 import { FaChartLine, FaEdit } from "react-icons/fa";
 import { Link } from "react-router-dom";
 import { FaStar, FaTrashAlt } from "react-icons/fa";
